# optimise defaultpreemption: enumerate fewer candidates