--- a/staging/src/k8s.io/kube-scheduler/config/v1beta1/register.go
+++ b/staging/src/k8s.io/kube-scheduler/config/v1beta1/register.go
@@ -38,6 +38,7 @@ var (
 func addKnownTypes(scheme *runtime.Scheme) error {
 	scheme.AddKnownTypes(SchemeGroupVersion,
 		&KubeSchedulerConfiguration{},
+		&DefaultPreemptionArgs{},
 		&InterPodAffinityArgs{},
 		&NodeLabelArgs{},
 		&NodeResourcesFitArgs{},