… pkg/scheduler/apis/config/v1beta1/zz_generated.conversion.go
… pkg/scheduler/apis/config/zz_generated.deepcopy.go
… pkg/scheduler/framework/plugins/defaultpreemption/default_preemption_test.go
… staging/src/k8s.io/kube-scheduler/config/v1beta1/zz_generated.deepcopy.go