<!--  Thanks for sending a pull request!  Here are some tips for you:

1. If this is your first time, please read our contributor guidelines: https://git.k8s.io/community/contributors/guide/first-contribution.md#your-first-contribution and developer guide https://git.k8s.io/community/contributors/devel/development.md#development-guide
2. Please label this pull request according to what type of issue you are addressing, especially if this is a release targeted pull request. For reference on required PR/issue labels, read here:
https://git.k8s.io/community/contributors/devel/sig-release/release.md#issuepr-kind-label
3. Ensure you have added or ran the appropriate tests for your PR: https://git.k8s.io/community/contributors/devel/sig-testing/testing.md
4. If you want *faster* PR reviews, read how: https://git.k8s.io/community/contributors/guide/pull-requests.md#best-practices-for-faster-reviews
5. If the PR is unfinished, see how to mark it: https://git.k8s.io/community/contributors/guide/pull-requests.md#marking-unfinished-pull-requests
-->

**What type of PR is this?**

/kind feature

**What this PR does / why we need it**: Instead of considering all nodes as preemption candidates, look for a smaller pool when dry running preemption and stop when a threshold number of candidates is found. In the average case (a good number of nodes are preemptible), this optimisation produces very good results (~25.3% throughput improvement). In the worst case (preemption is impossible), this approach will do no better or worse than the in-tree approach. The improvement is especially pronounced in large clusters (> 1k nodes). Clusters smaller than 100 nodes will not see any improvement from this optimisation.

The following benchmarks were performed with a 5,000 node cluster where 20,000 low priority pods are scheduled and then 5,000 high priority pods are scheduled (test run time is ~10 minutes). The high priority pods do not fit in any node without preempting one or more low priority pods. See the in-tree `scheduler_perf` integration benchmark for more details. Throughput percentile numbers are left out because median is 0 (kinda meaningless as a result).

| Quantity | Before | After | Change |
| - | - | - | - |
| Throughput average | 13.12 pods/sec | 16.45 pods/sec | +25.4% |
| `preemption_evaluation_seconds` average | 37.96ms | 30.20ms | -20.4% |
| `preemption_evaluation_seconds` p50 | 7.12ms | 6.22ms | -12.6% |
| `preemption_evaluation_seconds` p90 | 30.42ms | 24.01ms | -21.1% |
| `preemption_evaluation_seconds` p99 | 57.65ms | 41.75ms | -27.6% |
| Time spent in main thread (fraction of test time) | 6.18% | 2.51% | -59.3% |
| Time spent in parallel code (fraction of test time) | 16.26% | 6.25% | -61.6% |

Full diff (minus is after optimisation, plus is before): http://ix.io/2xHm

| Before | After |
| - | - |
| ![Screenshot_2020-09-15 scheduler_perf test cpu(2)](https://user-images.githubusercontent.com/7521600/93273385-b786d780-f785-11ea-9130-ee53617689fc.png) | ![Screenshot_2020-09-15 scheduler_perf test cpu(3)](https://user-images.githubusercontent.com/7521600/93273391-ba81c800-f785-11ea-8cdc-92e6d1eba0c8.png) |

No performance regression in 100 node tests (in fact, throughput is up slightly by 9%, not sure how).

**Which issue(s) this PR fixes**:
<!--
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related issues/tests in a comment and do not use `Fixes`_*
-->
Ref #89036 

**Special notes for your reviewer**:

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required:
Enter your extended release note in the block below. If the PR requires additional action from users switching to the new release, include the string "action required".

For more information on release notes see: https://git.k8s.io/community/contributors/guide/release-notes.md
-->
```release-note
Fewer candidates are enumerated for preemption to improve performance in large clusters
```

**Additional documentation e.g., KEPs (Kubernetes Enhancement Proposals), usage docs, etc.**:

<!--
This section can be blank if this pull request does not require a release note.

When adding links which point to resources within git repositories, like
KEPs or supporting documentation, please reference a specific commit and avoid
linking directly to the master branch. This ensures that links reference a
specific point in time, rather than a document that may change over time.

See here for guidance on getting permanent links to files: https://help.github.com/en/articles/getting-permanent-links-to-files

Please use the following format for linking documentation:
- [KEP]: <link>
- [Usage]: <link>
- [Other doc]: <link>
-->
```docs

```

/sig scheduling
/cc @alculquicondor @ahg-g