@adtac 2020-09-15
Will add/modify tests tomorrow.