@alculquicondor 2020-09-16
Did you add a new performance test case? Or are the above numbers from the existing ones?

Also, avoid `fixes #123` from the description, as that will cause github to close the issue once this merges.