@ahg-g 2020-09-16
Just noticed a problem with the windowing approach, it breaks the promise of picking "A node with minimum number of PDB violations.": https://github.com/kubernetes/kubernetes/blob/5d095c8a4427954049341ef8c0464545c03f6acf/pkg/scheduler/framework/plugins/defaultpreemption/default_preemption.go#L356

I am not sure we want to break that promise.