@alculquicondor 2020-09-16
> I am not sure we want to break that promise.

I think it's ok to break it as long as the default is kept at 100%