@ahg-g 2020-09-16
> > I am not sure we want to break that promise.
> 
> I think it's ok to break it as long as the default is kept at 100%

I think we can modify the promise in the following way:
1) the scheduler is free to select any node where the evicted pods don't violate pod disruption budget
2) if we can't find a node where the evicted pods don't violate pdb, then we have to select one where the evicted pods have the lowest impact on pdb