@Huang-Wei 2020-09-16
> I think we can modify the promise in the following way:
> 
> 1. the scheduler is free to select any node where the evicted pods don't violate pod disruption budget (the minimum within the window)
> 2. if we can't find a node where the evicted pods don't violate pdb, then we have to select one where the evicted pods have the lowest impact on pdb across all nodes

The semantics looks good to me, however, in practice, we should evaluate the difficulty of getting them implemented. If it's too much work (I'd assume we need to pre-know the nodes that don't have pods violating PDBs, so as to prioritize them in the searching scope), maybe it's ok to break the promise - as we're just increasing the disruption budget by 1, rather than totally breaking the PDB violation (semantics of PDB is guarded by API server).