@alculquicondor 2020-09-16
> The semantics looks good to me, however, in practice, we should evaluate the difficulty of getting them implemented. 

It shouldn't be hard at all. It's a matter of continuing iterating while:
- a node with zero violations is not found, OR
- percentage has not been reached.