@adtac 2020-09-16
Yes, it's just a matter of changing the context cancel condition slightly; we already have the PDB violation data for each node to make that decision. I'm still working on the component config changes haha, they're pretty confusing :)