@Huang-Wei 2020-09-16
> It shouldn't be hard at all. It's a matter of continuing iterating while:
> 
> * a node with zero violations is not found, OR
> * percentage has not been reached.

What if the percentage has been reached, and all so-far-calculated candidates have non-zero violations? Would you continue searching?