@adtac 2020-09-16
>What if the percentage has been reached, and all so-far-calculated candidates have non-zero violations? Would you continue searching?

Yes. The exit condition is basically `len(candidates) > threshold && nonViolatingNodeFound`. In the worst case (all nodes need to be searched), we don't do any better/worse than the in-tree approach. In the average case, we check fewer nodes.