@Huang-Wei 2020-09-16
> Yes. The exit condition is basically `len(candidates) > threshold && nonViolatingNodeFound`. In the worst case (all nodes need to be searched), we don't do any better/worse than the in-tree approach. In the average case, we check fewer nodes.

SG.