@adtac 2020-09-21
Unfortunately, github won't let me reply to comments until I post my pending comments, so I'll copy paste the important ones until I'm ready to post the whole set of comments (which I'll do once I push my changes -- currently writing tests).

>I feel the name can be simplified, like PercentageOfCandidateNodes?

technically, it's the percentage of nodes that are considered to be potential candidates, so "percentage of candidate nodes" would be slightly misleading (but not really since it's just yet another filtering phase within the plugin).

I'm open to the suggestion as it's simpler though :+1: