@Huang-Wei 2020-09-21
> I'm open to the suggestion as it's simpler though

@alculquicondor 's suggestion (`MaxCandidateNodesPercentage`) looks good to me.

(btw: there are no much points we can refer across k8s codebase, but I do find a name `CurrentCPUUtilizationPercentage` in autoscaler, so I think keeps the "units" at the end is more conventional, which is also aligned with prometheus metric names)