@adtac 2020-10-07
/test pull-kubernetes-e2e-kind